refactor: Fixed imports after having created utils package
