+++ Engine/Utils/Maths/Transform.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Numerics;
 
-namespace AntEngine.Maths
+namespace AntEngine.Utils.Maths
 {
     /// <summary>
     /// Contains the data needed for spatial manipulation of entities.
