+++ Tests/Engine/PerceptionMapTests.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Numerics;
-using AntEngine.Maths;
+using AntEngine.Utils.Maths;
 using Xunit;
 
 namespace Tests.Engine
